# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Web service API in Go using Gin framework.",
        "title": "Collagen Web Service",
        "termsOfService": "https://tos.santoshk.dev",
        "contact": {
            "name": "M Nurilman Baehaqi",
            "url": "https://twitter.com/MOXSPOY",
            "email": "mnurilmanbaehaqi@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
...
        "/user-detail/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Usually this endpoint used to process kyc",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user's selfie image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Api Key",
                        "name": "api_key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "User image that will be saved to the database_model",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_response_model.SuccessAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_response_model.SuccessAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}